backend/requirements.txt
asgiref==3.8.1
Django==5.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
gunicorn
djangorestframework-simplejwt
celery
redis==5.0.3